Unified dialect selection in libyul/SyntaxTest and YulOptimizerTest.
