+++ test/libyul/Common.cpp
@@ -31,6 +31,7 @@
 #include <libyul/AsmPrinter.h>
 #include <libyul/AssemblyStack.h>
 #include <libyul/backends/evm/EVMDialect.h>
+#include <libyul/backends/wasm/WasmDialect.h>
 
 #include <liblangutil/Scanner.h>
 #include <liblangutil/ErrorReporter.h>
@@ -105,4 +106,52 @@ yul::Block yul::test::disambiguate(string const& _source, bool _yul)
 string yul::test::format(string const& _source, bool _yul)
 {
 	return yul::AsmPrinter()(*parse(_source, _yul).first);
+}
+
+namespace
+{
+std::map<string const, yul::Dialect const& (*)(langutil::EVMVersion)> const validDialects = {
+	{
+		"evm",
+		[](langutil::EVMVersion _evmVersion) -> yul::Dialect const&
+		{ return yul::EVMDialect::strictAssemblyForEVMObjects(_evmVersion); }
+	},
+	{
+		"evmTyped",
+		[](langutil::EVMVersion _evmVersion) -> yul::Dialect const&
+		{ return yul::EVMDialectTyped::instance(_evmVersion); }
+	},
+	{
+		"yul",
+		[](langutil::EVMVersion) -> yul::Dialect const&
+		{ return yul::Dialect::yulDeprecated(); }
+	},
+	{
+		"ewasm",
+		[](langutil::EVMVersion) -> yul::Dialect const&
+		{ return yul::WasmDialect::instance(); }
+	}
+};
+
+vector<string> validDialectNames()
+{
+	vector<string> names{size(validDialects), ""};
+	transform(begin(validDialects), end(validDialects), names.begin(), [](auto const& dialect) { return dialect.first; });
+
+	return names;
+}
+}
+
+yul::Dialect const& yul::test::dialect(std::string const& _name, langutil::EVMVersion _evmVersion)
+{
+	if (!validDialects.count(_name))
+		BOOST_THROW_EXCEPTION(runtime_error{
+			"Invalid Dialect \"" +
+			_name +
+			"\". Valid dialects are " +
+			util::joinHumanReadable(validDialectNames(), ", ", " and ") +
+			"."
+		});
+
+	return validDialects.at(_name)(_evmVersion);
 }
