+++ test/libyul/SyntaxTest.cpp
@@ -19,14 +19,13 @@
 #include <libyul/AsmAnalysisInfo.h>
 #include <libyul/AsmParser.h>
 
-#include <libyul/backends/wasm/WasmDialect.h>
-#include <libyul/backends/evm/EVMDialect.h>
-#include <libyul/backends/wasm/WasmDialect.h>
-
 #include <liblangutil/EVMVersion.h>
 #include <liblangutil/Exceptions.h>
 
+#include <test/libyul/Common.h>
 #include <test/libyul/SyntaxTest.h>
+
+#include <test/Common.h>
 
 using namespace std;
 using namespace solidity;
@@ -34,44 +33,8 @@ using namespace solidity::util;
 using namespace solidity::langutil;
 using namespace solidity::yul::test;
 
-namespace
-{
-std::map<string const, yul::Dialect const& (*)(langutil::EVMVersion)> const validDialects = {
-	{
-		"evm",
-		[](langutil::EVMVersion _evmVersion) -> yul::Dialect const&
-		{ return yul::EVMDialect::strictAssemblyForEVM(_evmVersion); }
-	},
-	{
-		"evmTyped",
-		[](langutil::EVMVersion _evmVersion) -> yul::Dialect const&
-		{ return yul::EVMDialectTyped::instance(_evmVersion); }
-	},
-	{
-		"yul",
-		[](langutil::EVMVersion) -> yul::Dialect const&
-		{ return yul::Dialect::yulDeprecated(); }
-	},
-	{
-		"ewasm",
-		[](langutil::EVMVersion) -> yul::Dialect const&
-		{ return yul::WasmDialect::instance(); }
-	}
-};
-
-vector<string> validDialectNames()
-{
-	vector<string> names{size(validDialects), ""};
-	transform(begin(validDialects), end(validDialects), names.begin(), [](auto const& dialect) { return dialect.first; });
-
-	return names;
-}
-}
-
 void SyntaxTest::parseAndAnalyze()
 {
-	yul::Dialect const& dialect = validDialects.at(m_dialectName)(m_evmVersion);
-
 	if (m_sources.size() != 1)
 		BOOST_THROW_EXCEPTION(runtime_error{"Expected only one source for yul test."});
 
@@ -82,12 +45,12 @@ void SyntaxTest::parseAndAnalyze()
 	ErrorReporter errorReporter{errorList};
 
 	auto scanner = make_shared<Scanner>(CharStream(source, name));
-	auto parserResult = yul::Parser(errorReporter, dialect).parse(scanner, false);
+	auto parserResult = yul::Parser(errorReporter, *m_dialect).parse(scanner, false);
 
 	if (parserResult)
 	{
 		yul::AsmAnalysisInfo analysisInfo;
-		yul::AsmAnalyzer(analysisInfo, errorReporter, dialect).analyze(*parserResult);
+		yul::AsmAnalyzer(analysisInfo, errorReporter, *m_dialect).analyze(*parserResult);
 	}
 
 	for (auto const& error: errorList)
@@ -115,14 +78,6 @@ void SyntaxTest::parseAndAnalyze()
 SyntaxTest::SyntaxTest(string const& _filename, langutil::EVMVersion _evmVersion):
 	CommonSyntaxTest(_filename, _evmVersion)
 {
-	m_dialectName = m_reader.stringSetting("dialect", "evmTyped");
-
-	if (!validDialects.count(m_dialectName))
-		BOOST_THROW_EXCEPTION(runtime_error{
-			"Invalid Dialect \"" +
-			m_dialectName +
-			"\". Valid dialects are " +
-			joinHumanReadable(validDialectNames(), ", ", " and ") +
-			"."
-		});
+	string dialectName = m_reader.stringSetting("dialect", "evmTyped");
+	m_dialect = &dialect(dialectName, solidity::test::CommonOptions::get().evmVersion());
 }
